+++ src/firecracker/src/api_server/request/snapshot.rs
@@ -98,6 +98,7 @@ fn parse_put_snapshot_load(body: &Body) -> Result<ParsedRequest, RequestError> {
                 // either `mem_file_path` or `mem_backend` field is always specified.
                 backend_path: snapshot_config.mem_file_path.unwrap(),
                 backend_type: MemBackendType::File,
+                shared: false,
             }
         }
     };
@@ -185,6 +186,7 @@ mod tests {
             mem_backend: MemBackendConfig {
                 backend_path: PathBuf::from("bar"),
                 backend_type: MemBackendType::File,
+                shared: false,
             },
             track_dirty_pages: false,
             resume_vm: false,
@@ -217,6 +219,7 @@ mod tests {
             mem_backend: MemBackendConfig {
                 backend_path: PathBuf::from("bar"),
                 backend_type: MemBackendType::File,
+                shared: false,
             },
             track_dirty_pages: true,
             resume_vm: false,
@@ -249,6 +252,7 @@ mod tests {
             mem_backend: MemBackendConfig {
                 backend_path: PathBuf::from("bar"),
                 backend_type: MemBackendType::Uffd,
+                shared: false,
             },
             track_dirty_pages: false,
             resume_vm: true,
@@ -287,6 +291,7 @@ mod tests {
             mem_backend: MemBackendConfig {
                 backend_path: PathBuf::from("bar"),
                 backend_type: MemBackendType::Uffd,
+                shared: false,
             },
             track_dirty_pages: false,
             resume_vm: true,
@@ -319,6 +324,7 @@ mod tests {
             mem_backend: MemBackendConfig {
                 backend_path: PathBuf::from("bar"),
                 backend_type: MemBackendType::File,
+                shared: false,
             },
             track_dirty_pages: false,
             resume_vm: true,