+++ src/vmm/src/rpc_interface.rs
@@ -1325,6 +1325,7 @@ mod tests {
                 mem_backend: MemBackendConfig {
                     backend_type: MemBackendType::File,
                     backend_path: PathBuf::new(),
+                    shared: false,
                 },
                 track_dirty_pages: false,
                 resume_vm: false,