+++ src/vmm/tests/integration_tests.rs
@@ -311,6 +311,7 @@ fn verify_load_snapshot(snapshot_file: TempFile, memory_file: TempFile) {
             mem_backend: MemBackendConfig {
                 backend_path: memory_file.as_path().to_path_buf(),
                 backend_type: MemBackendType::File,
+                shared: false,
             },
             track_dirty_pages: false,
             resume_vm: true,
@@ -397,6 +398,7 @@ fn verify_load_snap_disallowed_after_boot_resources(res: VmmAction, res_name: &s
         mem_backend: MemBackendConfig {
             backend_path: memory_file.as_path().to_path_buf(),
             backend_type: MemBackendType::File,
+            shared: false,
         },
         track_dirty_pages: false,
         resume_vm: false,